+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Return string symbols instead of objects when requesting the account symbols.
+* Remove export of spread type.
 * Remove spread type property from symbols.
 * Expose puppeteer page instance in browser tab.
 * Define monthly timeframe type.